backend/requirements.txt
fastapi
uvicorn
python-socketio
openai
httpx
python-dotenv
websockets
slowapi
